Добавил usb драйвер(хз как он работает) теперь ввод с клавиатуры поддерживает юсб. Также теперь система 64битная

+++ stdlib.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void* itoa(int value, char* buffer, int base);
+void* itoa(long value, char* buffer, int base);
